+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/member/service/MemberService.java
@@ -2,10 +2,28 @@ package com.nextdoor.nextdoor.domain.member.service;
 
 import com.nextdoor.nextdoor.domain.member.controller.dto.request.MemberExtraInfoSaveRequestDto;
 import com.nextdoor.nextdoor.domain.member.controller.dto.response.MemberResponseDto;
+import com.nextdoor.nextdoor.domain.member.domain.model.Member;
+import com.nextdoor.nextdoor.domain.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface MemberService {
+@RequiredArgsConstructor
+@Service
+@Transactional
+public class MemberService {
 
-    MemberResponseDto updateMember(Long memberId, MemberExtraInfoSaveRequestDto memberDto);
+    private final MemberRepository memberRepository;
 
-    MemberResponseDto retrieveMember(Long memberId);
+    public MemberResponseDto updateMember(Long memberId, MemberExtraInfoSaveRequestDto memberDto) {
+        Member member = memberRepository.findById(memberId).orElseThrow();
+        member.updateBirth(memberDto.getBirth());
+        member.updateGender(memberDto.getGender());
+        member.updateAddress(memberDto.getAddress());
+        return MemberResponseDto.from(member);
+    }
+
+    public MemberResponseDto retrieveMember(Long memberId) {
+        return MemberResponseDto.from(memberRepository.findById(memberId).orElseThrow());
+    }
 }
